requirements.txt
# Core runtime dependencies
beautifulsoup4>=4.9.3
PyYAML>=5.4.1
aiohttp>=3.8.1
jinja2>=3.0.1
colorlog>=6.0.0
reportlab>=3.6.8
